Polished up some ring gun and basic gun
Made it so that ring and  gun projectiles collide and dequeue successfully when hitting the environment or an opponent. Changed up basic gun physics so that it shoots straight in the direction the gun points to when used on walking enemies(when walking enemies turn, the firing direction gets messed up from the flip as the rotation suddenly shifts by 180 degrees). Made a muzzle so the bullets don't appear from the center of the gun(the rings don't shoot of from the middle of the barrel).